+++ app/src/main/java/com/pbs/expenseApp/data/converters/Converters.kt
@@ -2,17 +2,21 @@ package com.pbs.expenseApp.data.converters
 
 import androidx.room.ProvidedTypeConverter
 import androidx.room.TypeConverter
-import java.util.Date
+import java.time.Instant
+import java.time.LocalDate
+import java.time.ZoneId
 
 @ProvidedTypeConverter
 class DateConverters {
+
+    private val zoneId: ZoneId = ZoneId.systemDefault()
     @TypeConverter
-    fun fromTimestamp(value: Long?): Date? {
-        return value?.let { Date(it) }
+    fun fromTimestamp(value: Long?): LocalDate? {
+        return value?.let { Instant.ofEpochSecond(it).atZone(zoneId).toLocalDate() }
     }
 
     @TypeConverter
-    fun dateToTimestamp(date: Date?): Long? {
-        return date?.time?.toLong()
+    fun dateToTimestamp(date: LocalDate?): Long? {
+        return date?.atStartOfDay(zoneId)?.toEpochSecond()
     }
 }