+++ app/src/main/java/com/pbs/expenseApp/domain/model/User.kt
@@ -2,10 +2,15 @@ package com.pbs.expenseApp.domain.model
 
 import androidx.room.Entity
 import androidx.room.PrimaryKey
+import androidx.room.TypeConverters
+import com.pbs.expenseApp.data.converters.DateConverters
+import java.time.LocalDate
 
 @Entity(tableName = "users")
-data class User(
+@TypeConverters(DateConverters::class)
+data class User (
     @PrimaryKey()
-    val uuid: String,
-    val monthlySavings: Int
+    val uuid: String = "",
+    var monthlySavings: Int = 0,
+    val creationDate: LocalDate = LocalDate.now()
 )