+++ app/src/main/java/com/pbs/expenseApp/ui/AppViewModelProvider.kt
@@ -7,6 +7,7 @@ import androidx.lifecycle.viewmodel.viewModelFactory
 import com.pbs.expenseApp.ExpenseApplication
 import com.pbs.expenseApp.ui.viewmodels.CategoryViewModel
 import com.pbs.expenseApp.ui.viewmodels.ConfigurationViewModel
+import com.pbs.expenseApp.ui.viewmodels.HorizontalPagerViewModel
 import com.pbs.expenseApp.ui.viewmodels.ExpenseViewModel
 import com.pbs.expenseApp.ui.viewmodels.ExposedDropDownViewModel
 import com.pbs.expenseApp.ui.viewmodels.UserViewModel
@@ -33,6 +34,9 @@ object AppViewModelProvider {
                 expenseApplication().container.expenseRepository,
                 expenseApplication().baseContext
             )
+        }
+        initializer {
+            HorizontalPagerViewModel()
         }
     }
 }