+++ app/src/main/java/com/pbs/expenseApp/ui/screens/AddExpense.kt
@@ -1,5 +1,7 @@
 package com.pbs.expenseApp.ui.screens
 
+import androidx.compose.foundation.ExperimentalFoundationApi
+import androidx.compose.foundation.layout.fillMaxHeight
 import androidx.compose.foundation.layout.fillMaxSize
 import androidx.compose.foundation.layout.fillMaxWidth
 import androidx.compose.foundation.layout.height
@@ -15,25 +17,32 @@ import com.pbs.expenseApp.R
 import com.pbs.expenseApp.ui.AppViewModelProvider
 import com.pbs.expenseApp.ui.components.AppCard
 import com.pbs.expenseApp.ui.components.AppColumn
+import com.pbs.expenseApp.ui.components.AppHorizontalPager
 import com.pbs.expenseApp.ui.components.AppRow
+import com.pbs.expenseApp.ui.components.AppText
 import com.pbs.expenseApp.ui.composables.MyMonthlySavingText
+import com.pbs.expenseApp.ui.viewmodels.HorizontalPagerViewModel
 import com.pbs.expenseApp.ui.viewmodels.UserViewModel
 import com.pbs.expenseApp.utils.AppUtils
 import kotlinx.coroutines.launch
 
+@OptIn(ExperimentalFoundationApi::class)
 @Composable
 fun AddExpense() {
     val context = LocalContext.current
     val userVM: UserViewModel = viewModel(factory = AppViewModelProvider.Factory)
+    val pagerVM: HorizontalPagerViewModel = viewModel(factory = AppViewModelProvider.Factory)
+
+    pagerVM.calculatePageCount(userVM.user.creationDate)
 
     LaunchedEffect(key1 = true) {
-        userVM.viewModelScope.launch { userVM.getMonthlySavings(AppUtils.getAppId(context)) }
+        userVM.viewModelScope.launch { userVM.getUser(AppUtils.getAppId(context)) }
     }
 
     AppColumn(
         modifier = Modifier
-        .fillMaxSize()
-        .padding(dimensionResource(id = R.dimen.padding_sm))
+            .fillMaxWidth()
+            .padding(dimensionResource(id = R.dimen.padding_sm))
     ) {
         AppCard(modifier = Modifier
             .fillMaxWidth()
@@ -41,9 +50,30 @@ fun AddExpense() {
         ) {
             AppRow(modifier = Modifier
                 .fillMaxSize()
-                .padding(start = dimensionResource(id = R.dimen.padding_sm), end = dimensionResource(id = R.dimen.padding_sm))
+                .padding(
+                    start = dimensionResource(id = R.dimen.padding_sm),
+                    end = dimensionResource(id = R.dimen.padding_sm)
+                )
             ) {
-                MyMonthlySavingText(monthlySavings = userVM.monthlySavings)
+                MyMonthlySavingText(monthlySavings = userVM.user.monthlySavings)
+            }
+        }
+
+        AppRow(modifier = Modifier
+            .fillMaxWidth()
+            .height(dimensionResource(R.dimen.card_height_sm))
+        ) {
+            AppText(text = pagerVM.formattedPagerDate)
+        }
+
+        AppColumn(modifier = Modifier
+            .fillMaxSize()
+        ) {
+            AppHorizontalPager(pageCount = pagerVM.pageCount) {
+                AppText(
+                    text = "Page: $it",
+                    modifier = Modifier.fillMaxWidth().fillMaxHeight()
+                )
             }
         }
     }